+++ lib/libssh/src/packet_cb.c
@@ -64,10 +64,10 @@ SSH_PACKET_CALLBACK(ssh_packet_disconnect_callback){
     SSH_STRING_FREE(error_s);
   }
   SSH_LOG(SSH_LOG_PACKET, "Received SSH_MSG_DISCONNECT %d:%s",
-                          code, error != NULL ? error : "no error");
+                          (int)code, error != NULL ? error : "no error");
   ssh_set_error(session, SSH_FATAL,
       "Received SSH_MSG_DISCONNECT: %d:%s",
-      code, error != NULL ? error : "no error");
+      (int)code, error != NULL ? error : "no error");
   SAFE_FREE(error);
 
   ssh_socket_close(session->socket);
@@ -218,7 +218,7 @@ SSH_PACKET_CALLBACK(ssh_packet_ext_info)
         return SSH_PACKET_USED;
     }
 
-    SSH_LOG(SSH_LOG_PACKET, "Follows %u extensions", nr_extensions);
+    SSH_LOG(SSH_LOG_PACKET, "Follows %u extensions", (unsigned)nr_extensions);
 
     for (i = 0; i < nr_extensions; i++) {
         char *name = NULL;
